% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithmic Leveraging for linear regression.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Algorithmic Leveraging for linear linear regression}
\usage{
algo_leverage(X, r)
}
\arguments{
\item{X}{input vector}

\item{r}{a list of subsample}
}
\value{
simulated sampled based on the value of leverage
}
\description{
Algorithmic Leveraging for linear linear regression
}
