% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Elastic Corrdinate Descent.R
\name{elnet_coord}
\alias{elnet_coord}
\title{Elastic Coordiante Descent Algorithm}
\usage{
elnet_coord(lambda, X, Y, alpha, maxit = 1e+05)
}
\arguments{
\item{lambda}{penalty term}

\item{X}{input dataset}

\item{Y}{output dataset}

\item{alpha}{constant multiplying penalty term}

\item{maxit}{maxium number of iterations}
}
\value{
a sequence of coefficients beta by fitting X to Y
}
\description{
Elastic Coordiante Descent Algorithm
}
