% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss Seidel and Jacobi.R
\name{solve_ols}
\alias{solve_ols}
\title{Gauss Seidel and Jacobi Algorithm for solving linear system of equations}
\usage{
solve_ols(A, b, c = 2, maxit = 10000, type = "GS")
}
\arguments{
\item{A}{matrix input}

\item{b}{output vector}

\item{type}{specify method's type, "GS" denotes "Gauss-Seidel", "Jc" represents Jacobi,"PJC" denotes Parallel Jacobi}

\item{n}{number of cores users can specify}

\item{maxiter}{number of iterations}
}
\value{
a vector of length n by solving AX=b
}
\description{
Gauss Seidel and Jacobi Algorithm for solving linear system of equations
}
